+++ sketch-compose-core/src/commonMain/kotlin/com/github/panpf/sketch/painter/ResizePainter.kt
@@ -20,15 +20,20 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.RememberObserver
 import androidx.compose.runtime.Stable
 import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.geometry.isUnspecified
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.drawscope.DrawScope
-import androidx.compose.ui.graphics.drawscope.inset
+import androidx.compose.ui.graphics.drawscope.translate
 import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.layout.times
 import com.github.panpf.sketch.resize.Scale
 import com.github.panpf.sketch.resize.Scale.CENTER_CROP
-import com.github.panpf.sketch.util.calculateScaleMultiplierWithFit
+import com.github.panpf.sketch.util.floatAlign
+import com.github.panpf.sketch.util.fromScale
+import com.github.panpf.sketch.util.name
 
 /**
  * Create a [ResizePainter] and remember it
@@ -36,6 +41,24 @@ import com.github.panpf.sketch.util.calculateScaleMultiplierWithFit
  * @see com.github.panpf.sketch.compose.core.common.test.painter.ResizePainterTest.testRememberResizePainter
  */
 @Composable
+fun rememberResizePainter(
+    painter: Painter,
+    size: Size,
+    contentScale: ContentScale = ContentScale.Fit,
+    alignment: Alignment = Alignment.Center
+): ResizePainter {
+    return remember(painter, size, contentScale, alignment) {
+        painter.resize(size, contentScale, alignment)
+    }
+}
+
+/**
+ * Create a [ResizePainter] and remember it
+ *
+ * @see com.github.panpf.sketch.compose.core.common.test.painter.ResizePainterTest.testRememberResizePainter
+ */
+@Composable
+@Deprecated("Use rememberResizePainter(painter, size, contentScale, alignment) instead")
 fun rememberResizePainter(painter: Painter, size: Size, scale: Scale = CENTER_CROP): ResizePainter {
     return remember(painter, size, scale) {
         painter.resize(size, scale)
@@ -47,6 +70,24 @@ fun rememberResizePainter(painter: Painter, size: Size, scale: Scale = CENTER_CR
  *
  * @see com.github.panpf.sketch.compose.core.common.test.painter.ResizePainterTest.testPainterResize
  */
+fun Painter.resize(
+    size: Size,
+    contentScale: ContentScale = ContentScale.Fit,
+    alignment: Alignment = Alignment.Center
+): ResizePainter {
+    return if (this is AnimatablePainter) {
+        ResizeAnimatablePainter(this, size, contentScale, alignment)
+    } else {
+        ResizePainter(this, size, contentScale, alignment)
+    }
+}
+
+/**
+ * Resize the [Painter] to the specified size
+ *
+ * @see com.github.panpf.sketch.compose.core.common.test.painter.ResizePainterTest.testPainterResize
+ */
+@Deprecated("Use resize(size, contentScale, alignment) instead")
 fun Painter.resize(size: Size, scale: Scale = CENTER_CROP): ResizePainter {
     return if (this is AnimatablePainter) {
         ResizeAnimatablePainter(this, size, scale)
@@ -61,17 +102,29 @@ fun Painter.resize(size: Size, scale: Scale = CENTER_CROP): ResizePainter {
  * @see com.github.panpf.sketch.compose.core.common.test.painter.ResizePainterTest
  */
 @Stable
-// TODO Support ContentScale and Alignment
-open class ResizePainter(
+open class ResizePainter constructor(
     val painter: Painter,
     val size: Size,
-    val scale: Scale
+    val contentScale: ContentScale = ContentScale.Fit,
+    val alignment: Alignment = Alignment.Center,
 ) : Painter(), RememberObserver, SketchPainter {
 
     override val intrinsicSize: Size = size
 
     private var alpha: Float = 1.0f
     private var colorFilter: ColorFilter? = null
+
+    @Deprecated(message = "Use ResizePainter(painter, size, contentScale, alignment) instead")
+    constructor(
+        painter: Painter,
+        size: Size,
+        scale: Scale,
+    ) : this(
+        painter = painter,
+        size = size,
+        contentScale = fromScale(scale).first,
+        alignment = fromScale(scale).second
+    )
 
     override fun onRemembered() {
         (painter as? RememberObserver)?.onRemembered()
@@ -99,17 +152,13 @@ open class ResizePainter(
         with(painter) {
             val dstSize: Size = this@onDraw.size
             val srcSize: Size = this@with.intrinsicSize
-            val scaledSrcSize = computeScaledSize(srcSize, dstSize)
             if (dstSize.isUnspecified || dstSize.isEmpty()) {
-                draw(scaledSrcSize, alpha, colorFilter)
+                draw(srcSize, alpha, colorFilter)
             } else {
-                val (horizontal, vertical) = when (scale) {
-                    Scale.START_CROP -> 0f to 0f
-                    Scale.END_CROP -> dstSize.width - scaledSrcSize.width to dstSize.height - scaledSrcSize.height
-                    else -> (dstSize.width - scaledSrcSize.width) / 2f to (dstSize.height - scaledSrcSize.height) / 2f
-                }
-                inset(horizontal = horizontal, vertical = vertical) {
-                    draw(scaledSrcSize, alpha, colorFilter)
+                val drawSize = computeScaledSize(srcSize, dstSize)
+                val offset = alignment.floatAlign(size = drawSize, space = dstSize)
+                translate(left = offset.x, top = offset.y) {
+                    draw(size = drawSize, alpha = alpha, colorFilter = colorFilter)
                 }
             }
         }
@@ -118,17 +167,9 @@ open class ResizePainter(
     private fun computeScaledSize(srcSize: Size, dstSize: Size): Size {
         if (srcSize.isUnspecified || srcSize.isEmpty()) return dstSize
         if (dstSize.isUnspecified || dstSize.isEmpty()) return dstSize
-        val sizeMultiplier = calculateScaleMultiplierWithFit(
-            srcWidth = srcSize.width,
-            srcHeight = srcSize.height,
-            dstWidth = dstSize.width,
-            dstHeight = dstSize.height,
-            fitScale = false
-        )
-        return Size(
-            width = srcSize.width * sizeMultiplier,
-            height = srcSize.height * sizeMultiplier
-        )
+        val sizeMultiplier = contentScale
+            .computeScaleFactor(srcSize = srcSize, dstSize = dstSize)
+        return srcSize * sizeMultiplier
     }
 
     override fun equals(other: Any?): Boolean {
@@ -137,7 +178,8 @@ open class ResizePainter(
         other as ResizePainter
         if (painter != other.painter) return false
         if (size != other.size) return false
-        return scale == other.scale
+        if (contentScale != other.contentScale) return false
+        return alignment == other.alignment
     }
 
     override fun hashCode(): Int {
@@ -145,11 +187,14 @@ open class ResizePainter(
         // Because size are value classes, they will be replaced by long.
         // Long will lose precision when converting hashCode, causing the hashCode generated by different srcOffset and srcSize to be the same.
         result = 31 * result + size.toString().hashCode()
-        result = 31 * result + scale.hashCode()
+        result = 31 * result + contentScale.hashCode()
+        result = 31 * result + alignment.hashCode()
         return result
     }
 
-    override fun toString(): String {
-        return "ResizePainter(painter=${painter.toLogString()}, size=${size.width}x${size.height}, scale=$scale)"
-    }
+    override fun toString(): String = "ResizePainter(" +
+            "painter=${painter.toLogString()}, " +
+            "size=${size.width}x${size.height}, " +
+            "contentScale=${contentScale.name}, " +
+            "alignment=${alignment.name})"
 }