+++ sketch-compose-core/src/commonTest/kotlin/com/github/panpf/sketch/compose/core/common/test/painter/ResizeAnimatablePainterTest.kt
@@ -1,13 +1,13 @@
 package com.github.panpf.sketch.compose.core.common.test.painter
 
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.ColorPainter
+import androidx.compose.ui.layout.ContentScale
 import com.github.panpf.sketch.painter.ResizeAnimatablePainter
 import com.github.panpf.sketch.painter.asEquitable
 import com.github.panpf.sketch.painter.toLogString
-import com.github.panpf.sketch.resize.Scale.CENTER_CROP
-import com.github.panpf.sketch.resize.Scale.START_CROP
 import com.github.panpf.sketch.test.utils.TestAnimatablePainter
 import com.github.panpf.sketch.test.utils.block
 import kotlinx.coroutines.test.runTest
@@ -19,10 +19,12 @@ import kotlin.test.assertTrue
 
 class ResizeAnimatablePainterTest {
 
+    // TODO test constructor
+
     @Test
     fun testStartStop() = runTest {
         val animatablePainter = TestAnimatablePainter(ColorPainter(Color.Yellow))
-        val wrapper = ResizeAnimatablePainter(animatablePainter, Size(100f, 500f), CENTER_CROP)
+        val wrapper = ResizeAnimatablePainter(animatablePainter, Size(100f, 500f))
 
         assertFalse(actual = animatablePainter.isRunning())
         assertFalse(actual = wrapper.isRunning())
@@ -48,36 +50,41 @@ class ResizeAnimatablePainterTest {
         val element1 = ResizeAnimatablePainter(
             painter = TestAnimatablePainter(ColorPainter(Color.Red).asEquitable()),
             size = Size(100f, 500f),
-            scale = CENTER_CROP,
         )
         val element11 = ResizeAnimatablePainter(
             painter = TestAnimatablePainter(ColorPainter(Color.Red).asEquitable()),
             size = Size(100f, 500f),
-            scale = CENTER_CROP,
         )
         val element2 = ResizeAnimatablePainter(
             painter = TestAnimatablePainter(ColorPainter(Color.Green).asEquitable()),
             size = Size(100f, 500f),
-            scale = CENTER_CROP,
         )
         val element3 = ResizeAnimatablePainter(
             painter = TestAnimatablePainter(ColorPainter(Color.Red).asEquitable()),
             size = Size(500f, 100f),
-            scale = CENTER_CROP,
         )
         val element4 = ResizeAnimatablePainter(
             painter = TestAnimatablePainter(ColorPainter(Color.Red).asEquitable()),
             size = Size(100f, 500f),
-            scale = START_CROP,
+            contentScale = ContentScale.FillBounds,
+        )
+        val element5 = ResizeAnimatablePainter(
+            painter = TestAnimatablePainter(ColorPainter(Color.Red).asEquitable()),
+            size = Size(100f, 500f),
+            alignment = Alignment.BottomCenter,
         )
 
         assertEquals(element1, element11)
         assertNotEquals(element1, element2)
         assertNotEquals(element1, element3)
         assertNotEquals(element1, element4)
+        assertNotEquals(element1, element5)
         assertNotEquals(element2, element3)
         assertNotEquals(element2, element4)
+        assertNotEquals(element2, element5)
         assertNotEquals(element3, element4)
+        assertNotEquals(element3, element5)
+        assertNotEquals(element4, element5)
         assertNotEquals(element1, null as Any?)
         assertNotEquals(element1, Any())
 
@@ -85,9 +92,13 @@ class ResizeAnimatablePainterTest {
         assertNotEquals(element1.hashCode(), element2.hashCode())
         assertNotEquals(element1.hashCode(), element3.hashCode())
         assertNotEquals(element1.hashCode(), element4.hashCode())
+        assertNotEquals(element1.hashCode(), element5.hashCode())
         assertNotEquals(element2.hashCode(), element3.hashCode())
         assertNotEquals(element2.hashCode(), element4.hashCode())
+        assertNotEquals(element2.hashCode(), element5.hashCode())
         assertNotEquals(element3.hashCode(), element4.hashCode())
+        assertNotEquals(element3.hashCode(), element5.hashCode())
+        assertNotEquals(element4.hashCode(), element5.hashCode())
     }
 
     @Test
@@ -96,10 +107,9 @@ class ResizeAnimatablePainterTest {
         ResizeAnimatablePainter(
             painter = animatablePainter,
             size = Size(100f, 500f),
-            scale = CENTER_CROP
         ).apply {
             assertEquals(
-                expected = "ResizeAnimatablePainter(painter=${animatablePainter.toLogString()}, size=100.0x500.0, scale=CENTER_CROP)",
+                expected = "ResizeAnimatablePainter(painter=${animatablePainter.toLogString()}, size=100.0x500.0, contentScale=Fit, alignment=Center)",
                 actual = toString()
             )
         }